HBASE-25843 move master http-related code into o.a.h.h.master.http
