+++ hbase-server/src/main/java/org/apache/hadoop/hbase/master/http/RegionReplicaInfo.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.hadoop.hbase.master.webapp;
+package org.apache.hadoop.hbase.master.http;
 
 import java.util.Collections;
 import java.util.LinkedHashMap;
